feat: Add configurable $IPFS_EXEC
